move category select under wearable image
