+++ src/pages/add.tsx
@@ -102,7 +102,7 @@ export function AddPage() {
       <Form {...form}>
         <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-8">
           <FormItem>
-            <div className="grid grid-cols-2 gap-4">
+            <div className="flex gap-4">
               {wearablesFieldArray.fields.map((field, index) => (
                 <WearableAddCard
                   key={field.id}
